Add tests for Verify middleware
This adds the testing scenarios described in the functionality for the
method. The tests should cover most of the code surrounding the
functionality however, errors on the storage is not tested yet.

If the storage returns errors, this can be seen as undefined behavior
for now. However, in the case of errors, the middleware will fail and
return an InternalServerError.